Move deployment shared components
