+++ web/features/deployments/create-guide/ui/layout.tsx
@@ -5,7 +5,7 @@ import type { GuideStep } from '@/features/deployments/create-guide/state'
 import { cn } from '@langgenius/dify-ui/cn'
 import { ScrollArea } from '@langgenius/dify-ui/scroll-area'
 import { useTranslation } from 'react-i18next'
-import { TitleTooltip } from '@/features/deployments/components/title-tooltip'
+import { TitleTooltip } from '@/features/deployments/shared/components/title-tooltip'
 
 const GUIDE_PROGRESS_STEPS: GuideStep[] = ['source', 'release', 'target']
