+++ web/features/deployments/deploy-drawer/state/index.ts
@@ -11,21 +11,21 @@ import type {
   EnvVarBindingSlot,
   EnvVarValues,
   EnvVarValueSelection,
-} from '../../components/env-var-bindings'
-import type { RuntimeCredentialBindingSelections } from '../../components/runtime-credential-bindings-utils'
+} from '../../shared/components/env-var-bindings'
+import type { RuntimeCredentialBindingSelections } from '../../shared/components/runtime-credential-bindings-utils'
 import { EnvVarValueSource as ApiEnvVarValueSource } from '@dify/contracts/enterprise/types.gen'
 import { toast } from '@langgenius/dify-ui/toast'
 import { skipToken } from '@tanstack/react-query'
 import { atom } from 'jotai'
 import { atomWithMutation, atomWithQuery } from 'jotai-tanstack-query'
 import { consoleQuery } from '@/service/client'
-import { envVarBindingSlotFromContract } from '../../components/env-var-bindings-utils'
+import { envVarBindingSlotFromContract } from '../../shared/components/env-var-bindings-utils'
 import {
   hasMissingRequiredRuntimeCredentialBinding,
   runtimeCredentialSlotKey,
   selectedDeploymentRuntimeCredentials,
   selectedRuntimeCredentialSelections,
-} from '../../components/runtime-credential-bindings-utils'
+} from '../../shared/components/runtime-credential-bindings-utils'
 import { createDeploymentIdempotencyKey } from '../../shared/domain/idempotency'
 import { releaseDeploymentAction } from '../../shared/domain/release-action'
