+++ web/features/deployments/detail/access-tab/developer-api/section.tsx
@@ -13,8 +13,8 @@ import { useAtomValue } from 'jotai'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { consoleQuery } from '@/service/client'
-import { DeploymentEmptyState, DeploymentStateMessage } from '../../../components/empty-state'
 import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
+import { DeploymentEmptyState, DeploymentStateMessage } from '../../../shared/components/empty-state'
 import { CopyPill } from '../components/endpoint'
 import { developerApiSettingsQueryAtom } from '../state'
 import { ApiKeyGenerateMenu } from './api-key-generate-menu'
