+++ web/features/deployments/detail/access-tab/permissions/section.tsx
@@ -4,8 +4,8 @@ import type { EnvironmentAccessPolicy } from '@dify/contracts/enterprise/types.g
 import { useAtomValue } from 'jotai'
 import { useTranslation } from 'react-i18next'
 import { SkeletonRectangle } from '@/app/components/base/skeleton'
-import { DeploymentEmptyState, DeploymentStateMessage } from '../../../components/empty-state'
 import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
+import { DeploymentEmptyState, DeploymentStateMessage } from '../../../shared/components/empty-state'
 import { Section } from '../../components/section'
 import { accessSettingsQueryAtom } from '../state'
 import { EnvironmentPermissionRow } from './environment-permission-row'
