+++ web/features/deployments/detail/overview-tab/environment-tile.tsx
@@ -10,9 +10,9 @@ import { cn } from '@langgenius/dify-ui/cn'
 import { useAtomValue, useSetAtom } from 'jotai'
 import { useTranslation } from 'react-i18next'
 import Link from '@/next/link'
-import { TitleTooltip } from '../../components/title-tooltip'
 import { openDeployDrawerAtom } from '../../deploy-drawer/state'
 import { deploymentRouteAppInstanceIdAtom } from '../../route-state'
+import { TitleTooltip } from '../../shared/components/title-tooltip'
 import { releaseCommit } from '../../shared/domain/release'
 import { DeploymentStatusBadge } from '../../shared/ui/deployment-status-badge'
 import {
