+++ web/features/deployments/detail/overview-tab/index.tsx
@@ -3,8 +3,8 @@
 import { useAtomValue } from 'jotai'
 import { useTranslation } from 'react-i18next'
 import Link from '@/next/link'
-import { DeploymentStateMessage } from '../../components/empty-state'
 import { deploymentRouteAppInstanceIdAtom } from '../../route-state'
+import { DeploymentStateMessage } from '../../shared/components/empty-state'
 import { hasRuntimeInstanceDeployment } from '../../shared/domain/runtime-status'
 import { deploymentDetailOverviewQueryAtom } from '../state'
 import { AccessStatusSection, AccessStatusSectionSkeleton, ApiTokenSummarySection, ApiTokenSummarySectionSkeleton } from './access-status-section'
