+++ web/features/deployments/detail/overview-tab/release-hero.tsx
@@ -11,10 +11,10 @@ import { SkeletonRectangle } from '@/app/components/base/skeleton'
 import { useFormatTimeFromNow } from '@/hooks/use-format-time-from-now'
 import Link from '@/next/link'
 import { consoleQuery } from '@/service/client'
-import { DeploymentEmptyState } from '../../components/empty-state'
-import { TitleTooltip } from '../../components/title-tooltip'
 import { CreateReleaseControl } from '../../create-release'
 import { deploymentRouteAppInstanceIdAtom } from '../../route-state'
+import { DeploymentEmptyState } from '../../shared/components/empty-state'
+import { TitleTooltip } from '../../shared/components/title-tooltip'
 import { formatDate, releaseCommit } from '../../shared/domain/release'
 import { OVERVIEW_CARD_CLASS_NAME, OVERVIEW_ICON_CLASS_NAME } from './card-styles'
