+++ web/features/deployments/detail/versions-tab/deploy-release-menu.tsx
@@ -14,9 +14,9 @@ import { useAtomValue, useSetAtom } from 'jotai'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { consoleQuery } from '@/service/client'
-import { TitleTooltip } from '../../components/title-tooltip'
 import { openDeployDrawerAtom } from '../../deploy-drawer/state'
 import { deploymentRouteAppInstanceIdAtom } from '../../route-state'
+import { TitleTooltip } from '../../shared/components/title-tooltip'
 import { isUndeployedDeploymentRow } from '../../shared/domain/runtime-status'
 import { DETAIL_TABLE_ACTION_TRIGGER_CLASS_NAME } from '../components/detail-table-styles'
 import { DeleteReleaseDialog } from './delete-release-dialog'
